* 'develop' of https://github.com/TomHolvoet/SwarmView:
  Added zigzag trajectory, made Nerve use it (although not final), changed TinDrones such that they synchronize after going to their corkscrew position

Conflicts:
	trajectory/src/main/java/io/github/agentwise/swarmview/trajectory/rats/acts/introduction/NerveTrajectoryIntroduction.java

+++ trajectory/src/main/java/io/github/agentwise/swarmview/trajectory/rats/acts/attack/AttackTrajectories.java
@@ -38,7 +38,7 @@ public class AttackTrajectories {
 		this.attackRadius = attackRadius;
 	}
 
-	public FiniteTrajectory4d getTrajectory (int idDrone, int numberOfDrones, int sequenceOfAttack) {
+	public FiniteTrajectory4d getTrajectory (int idDrone, int numberOfDrones, double sequenceOfAttack) {
 
 		// first we create a line trajectory from the current position to the position to start the attack
